# Lammah-game-backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/games": {
      "post": {
        "operationId": "gamesCreate",
        "summary": "Create a new game with 2 teams",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameDto"
              },
              "examples": {
                "default": {
                  "summary": "Create game",
                  "value": {
                    "name": "Friday Family Game",
                    "teams": [
                      {
                        "name": "Team Falcons",
                        "members": [
                          "Muath",
                          "Sara"
                        ]
                      },
                      {
                        "name": "Team Stars",
                        "members": [
                          "Noura",
                          "Fahad"
                        ]
                      }
                    ],
                    "categoryIds": [
                      "66b8f5f2c9d7a8b1e3f00222",
                      "66b8f5f2c9d7a8b1e3f00223"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Game created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameCreationValidationErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Subscription required after free game is used",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "You need an active subscription to create more games.",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "gamesList",
        "summary": "Get all games",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Games retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}": {
      "get": {
        "operationId": "gamesGetById",
        "summary": "Get a specific game by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00444",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game retrieved successfully. Answers hidden by default.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Game with ID \"66b8f5f2c9d7a8b1e3f00444\" not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/questions/{gameQuestionId}": {
      "get": {
        "operationId": "gamesGetQuestionView",
        "summary": "Get one immutable game-question snapshot without its answer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gameQuestionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameQuestionViewEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Game or question not found"
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/questions/{gameQuestionId}/reveal": {
      "post": {
        "operationId": "gamesRevealQuestionView",
        "summary": "Reveal one game-question answer idempotently",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gameQuestionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameQuestionAnswerEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/questions/{gameQuestionId}/answer": {
      "get": {
        "operationId": "gamesGetQuestionAnswerView",
        "summary": "Get a previously revealed game-question answer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gameQuestionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameQuestionAnswerEnvelopeDto"
                }
              }
            }
          },
          "400": {
            "description": "Answer is not revealed"
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/questions/{gameQuestionId}/result": {
      "post": {
        "operationId": "gamesSubmitQuestionResult",
        "summary": "Finalize a game question for a team or no one",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gameQuestionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitGameQuestionResultDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid team, unrevealed answer, or already answered"
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/reveal-answer": {
      "post": {
        "operationId": "gamesRevealAnswer",
        "summary": "Reveal the correct answer for a question",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00444",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevealAnswerDto"
              },
              "examples": {
                "default": {
                  "summary": "Reveal answer",
                  "value": {
                    "questionId": "66b8f5f2c9d7a8b1e3f00333"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answer revealed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - question not found in game",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Question not found in this game board",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Game not found"
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/award-points": {
      "post": {
        "operationId": "gamesAwardPoints",
        "summary": "Award points to a team for answering correctly",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00444",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "questionId",
                  "teamIndex"
                ],
                "properties": {
                  "questionId": {
                    "type": "string",
                    "example": "66b8f5f2c9d7a8b1e3f00333"
                  },
                  "teamIndex": {
                    "type": "number",
                    "example": 0
                  }
                }
              },
              "examples": {
                "default": {
                  "summary": "Award points to team 0",
                  "value": {
                    "questionId": "66b8f5f2c9d7a8b1e3f00333",
                    "teamIndex": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Points awarded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid teamIndex or question already answered",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Question is already answered",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Game not found"
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/skip-question": {
      "post": {
        "operationId": "gamesSkipQuestion",
        "summary": "Skip a question without awarding points",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00444",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkipQuestionDto"
              },
              "examples": {
                "default": {
                  "summary": "Skip question",
                  "value": {
                    "questionId": "66b8f5f2c9d7a8b1e3f00333"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question skipped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - question already answered",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Question is already answered",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Game not found"
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/ranked-list-rounds/start": {
      "post": {
        "operationId": "gamesStartRankedListRound",
        "summary": "Start or resume a ranked-list round",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRankedListRoundDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedListRoundActionEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Lammah Game API",
    "description": "Runtime HTTP contract for Lammah game administration, content, and gameplay.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateGameDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "teams": {
            "minItems": 2,
            "maxItems": 2,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamDto"
            }
          },
          "categoryIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "teams",
          "categoryIds"
        ]
      },
...
      "GameMutationResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/GameResponseDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "data",
          "message"
        ]
      },
...
      "GameCreationValidationErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "code": {
            "type": "string",
            "example": "STANDARD_MISSING_200_QUESTIONS"
          },
          "message": {
            "type": "string"
          },
          "issueCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameCreationValidationDetailDto"
            }
          }
        },
        "required": [
          "statusCode",
          "error",
          "code",
          "message",
          "issueCodes",
          "details"
        ]
      },
...
    }
  }
}